Move alert/notification related code into seperated file
