+++ src/Makefile.omnicore.include
@@ -6,6 +6,7 @@ OMNICORE_H = \
   omnicore/errors.h \
   omnicore/log.h \
   omnicore/mdex.h \
+  omnicore/notifications.h \
   omnicore/omnicore.h \
   omnicore/parse_string.h \
   omnicore/pending.h \
@@ -28,6 +29,7 @@ OMNICORE_CPP = \
   omnicore/encoding.cpp \
   omnicore/log.cpp \
   omnicore/mdex.cpp \
+  omnicore/notifications.cpp \
   omnicore/omnicore.cpp \
   omnicore/parse_string.cpp \
   omnicore/pending.cpp \
