+++ src/omnicore/tx.cpp
@@ -6,6 +6,7 @@
 #include "omnicore/dex.h"
 #include "omnicore/log.h"
 #include "omnicore/mdex.h"
+#include "omnicore/notifications.h"
 #include "omnicore/omnicore.h"
 #include "omnicore/sp.h"
 #include "omnicore/sto.h"
